requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.3
cryptography==41.0.7
eth-account==0.11.0
aiosqlite==0.20.0
python-dotenv==1.0.1
